refactored MakeRecipe (separated functions into helpers) and fixed error where was not checking grams for isWeightUnit
